pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-examdiff"
version = "1.0.0"
description = "Professional file and directory comparison tool for Windows"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["diff", "comparison", "examdiff", "file-compare", "merge-tool"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Version Control",
    "Topic :: Text Editors :: Text Processing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
]

dependencies = [
    "customtkinter>=5.2.0",
    "tkinterdnd2>=0.3.0",
    "Pygments>=2.16.0",
    "chardet>=5.2.0",
    "diff-match-patch>=20230430",
    "watchdog>=3.0.0",
    "send2trash>=1.8.2",
    "reportlab>=4.0.5",
    "Pillow>=10.1.0",
    "markdown>=3.5.0",
    "PyYAML>=6.0.1",
    "GitPython>=3.1.40",
    "pywin32>=306; platform_system=='Windows'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "mypy>=1.7.0",
    "pyinstaller>=6.2.0",
]

[project.scripts]
examdiff = "main:main"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/python-examdiff"
Documentation = "https://github.com/YOUR_USERNAME/python-examdiff/wiki"
Repository = "https://github.com/YOUR_USERNAME/python-examdiff"
Issues = "https://github.com/YOUR_USERNAME/python-examdiff/issues"

[tool.setuptools]
packages = ["gui", "core", "utils", "plugins"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
